Correccion nombres en el servicio de reportes de companias y enums, tambien validacion del valor en el llamado el sp eliminacion de empresas

+++ UsaloYa.API/Controllers/CompanyController.cs
@@ -299,18 +299,26 @@ namespace UsaloYa.API.Controllers
         {
             try
             {
-                var requestor = await HeaderValidatorService.ValidateRequestorSameCompanyOrTopRol(RequestorId,  companyId, Role.Ventas, _dBContext);
-                return Unauthorized(AppConfig.NO_AUTORIZADO);
+                var requestor = await HeaderValidatorService.ValidateRequestorSameCompanyOrTopRol(
+                    RequestorId, companyId, Role.SysAdmin, _dBContext
+                );
+
+                if (requestor.UserId == 0)
+                    return Unauthorized(AppConfig.NO_AUTORIZADO);
+
+                if (days <= 0)
+                    return BadRequest(new { message = "$_Invalid_Value." });
 
                 var result = await _companyService.DeleteInactiveCompanies(days);
                 return Ok(result);
             }
             catch (Exception ex)
             {
-                _logger.LogError(ex, "GetMaxDaysToRefund.ApiError");
-                return StatusCode(500, new { message = "$_Excepcion_Ocurrida" });
+                _logger.LogError(ex, "DeleteInactiveCompanies.ApiError");
+                return StatusCode(500, new { message = "$_Database_Excepcion" });
             }
         }
+
 
     }
 }
